backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.8.0
python-multipart==0.0.20
httpx==0.28.1
redis==5.2.1
email-validator==2.2.0
bcrypt==4.1.3
reportlab==4.2.5
